deleted unnecessary comments and moved two files to the folder views
